requirements.txt
Django==3.2.8
django_multiselectfield==0.1.12
django_phonenumber_field==6.0.0
django_polymorphic==3.1.0
